Make tests into semantic tests
